refactor(modbus): split oversized view/dialog files
Behavior-preserving; pure code motion, log strings untouched. The three
~1300-line files became directories, lifting the widget-tree builder and
the layout render into sibling submodules (child modules reach the
parents' private validate/is_boolean_kind, so no API change):

- dialog/input.rs   -> input/{mod, build, render}
- dialog/selection.rs -> selection/{mod, build, render}
- view.rs -> view/{mod, mutate, overlay}: overlay holds PendingAction +
  the ModbusOverlay dispatch enum; mutate holds the async apply path
  (apply_add/apply_edit/delete/apply_setup/set_register_value +
  save_device_to/apply_order), pub(super) so mod.rs still drives them

Also clear the workspace clippy pass:
- modbus/server_core: allow too_many_arguments on the request helpers
  (intrinsic request-context args; bundling adds indirection)
- lua/ocpp: extract a Dispatched type alias for the mock-host field
- store/memory: contains_key over get().is_none()
- ocpp client handlers: drop() the intentionally-unawaited future
